add resource_efficiency, move some other functions to new file efficiency.py and create test cases
