+++ bct/algorithms/__init__.py
@@ -1,5 +1,6 @@
 from .centrality import *
 from .clustering import *
+from .efficiency import *
 from .generative import *
 from .modularity import *
 from .core import *
