+++ bct/algorithms/distance.py
@@ -1,6 +1,6 @@
 from __future__ import division, print_function
 import numpy as np
-from bct.utils import cuberoot, binarize, invert, BCTParamError, logtransform
+from bct.utils import binarize, BCTParamError, invert, logtransform
 from ..due import due, BibTeX
 from ..citations import LATORA2001, ONNELA2005, FAGIOLO2007, RUBINOV2010
 
@@ -467,217 +467,6 @@ def retrieve_shortest_path(s, t, hops, Pmat):
     return path
 
 
-@due.dcite(BibTeX(LATORA2001), description="Unweighted global efficiency")
-@due.dcite(BibTeX(ONNELA2005), description="Unweighted global efficiency")
-@due.dcite(BibTeX(FAGIOLO2007), description="Unweighted global efficiency")
-@due.dcite(BibTeX(RUBINOV2010), description="Unweighted global efficiency")
-def efficiency_bin(G, local=False):
-    '''
-    The global efficiency is the average of inverse shortest path length,
-    and is inversely related to the characteristic path length.
-
-    The local efficiency is the global efficiency computed on the
-    neighborhood of the node, and is related to the clustering coefficient.
-
-    Parameters
-    ----------
-    A : NxN np.ndarray
-        binary undirected connection matrix
-    local : bool
-        If True, computes local efficiency instead of global efficiency.
-        Default value = False.
-
-    Returns
-    -------
-    Eglob : float
-        global efficiency, only if local=False
-    Eloc : Nx1 np.ndarray
-        local efficiency, only if local=True
-    '''
-    def distance_inv(g):
-        D = np.eye(len(g))
-        n = 1
-        nPATH = g.copy()
-        L = (nPATH != 0)
-
-        while np.any(L):
-            D += n * L
-            n += 1
-            nPATH = np.dot(nPATH, g)
-            L = (nPATH != 0) * (D == 0)
-        D[np.logical_not(D)] = np.inf
-        D = 1 / D
-        np.fill_diagonal(D, 0)
-        return D
-
-    G = binarize(G)
-    n = len(G)  # number of nodes
-    if local:
-        E = np.zeros((n,))  # local efficiency
-
-        for u in range(n):
-            # V,=np.where(G[u,:])			#neighbors
-            # k=len(V)					#degree
-            # if k>=2:					#degree must be at least 2
-            #	e=distance_inv(G[V].T[V])
-            #	E[u]=np.sum(e)/(k*k-k)	#local efficiency computation
-
-            # find pairs of neighbors
-            V, = np.where(np.logical_or(G[u, :], G[u, :].T))
-            # inverse distance matrix
-            e = distance_inv(G[np.ix_(V, V)])
-            # symmetrized inverse distance matrix
-            se = e + e.T
-
-            # symmetrized adjacency vector
-            sa = G[u, V] + G[V, u].T
-            numer = np.sum(np.outer(sa.T, sa) * se) / 2
-            if numer != 0:
-                denom = np.sum(sa)**2 - np.sum(sa * sa)
-                E[u] = numer / denom  # local efficiency
-
-    else:
-        e = distance_inv(G)
-        E = np.sum(e) / (n * n - n)  # global efficiency
-    return E
-
-
-@due.dcite(BibTeX(LATORA2001), description="Weighted global efficiency")
-@due.dcite(BibTeX(ONNELA2005), description="Weighted global efficiency")
-@due.dcite(BibTeX(FAGIOLO2007), description="Weighted global efficiency")
-@due.dcite(BibTeX(RUBINOV2010), description="Weighted global efficiency")
-def efficiency_wei(Gw, local=False):
-    '''
-    The global efficiency is the average of inverse shortest path length,
-    and is inversely related to the characteristic path length.
-
-    The local efficiency is the global efficiency computed on the
-    neighborhood of the node, and is related to the clustering coefficient.
-
-    Parameters
-    ----------
-    W : NxN np.ndarray
-        undirected weighted connection matrix
-        (all weights in W must be between 0 and 1)
-    local = bool or enum
-        If True or 'local', computes local efficiency instead of global efficiency.
-        If False or 'global', uses the global efficiency
-        If 'original', will use the original algorithm provided by (Rubinov
-        & Sporns 2010). This version is not recommended. The local efficiency
-        calculation was improved in (Wang et al. 2016) as a true generalization
-        of the binary variant.
-        
-    Returns
-    -------
-    Eglob : float
-        global efficiency, only if local in (False, 'global')
-    Eloc : Nx1 np.ndarray
-        local efficiency, only if local in (True, 'local', 'original')
-
-    Notes
-    -----
-       The  efficiency is computed using an auxiliary connection-length
-    matrix L, defined as L_ij = 1/W_ij for all nonzero L_ij; This has an
-    intuitive interpretation, as higher connection weights intuitively
-    correspond to shorter lengths.
-       The weighted local efficiency broadly parallels the weighted
-    clustering coefficient of Onnela et al. (2005) and distinguishes the
-    influence of different paths based on connection weights of the
-    corresponding neighbors to the node in question. In other words, a path
-    between two neighbors with strong connections to the node in question
-    contributes more to the local efficiency than a path between two weakly
-    connected neighbors. Note that this weighted variant of the local
-    efficiency is hence not a strict generalization of the binary variant.
-
-    Algorithm:  Dijkstra's algorithm
-    '''
-    if local not in (True, False, 'local', 'global', 'original'):
-        raise BCTParamError("local param must be any of True, False, "
-            "'local', 'global', or 'original'")
-
-    def distance_inv_wei(G):
-        n = len(G)
-        D = np.zeros((n, n))  # distance matrix
-        D[np.logical_not(np.eye(n))] = np.inf
-
-        for u in range(n):
-            # distance permanence (true is temporary)
-            S = np.ones((n,), dtype=bool)
-            G1 = G.copy()
-            V = [u]
-            while True:
-                S[V] = 0  # distance u->V is now permanent
-                G1[:, V] = 0  # no in-edges as already shortest
-                for v in V:
-                    W, = np.where(G1[v, :])  # neighbors of smallest nodes
-                    td = np.array(
-                        [D[u, W].flatten(), (D[u, v] + G1[v, W]).flatten()])
-                    D[u, W] = np.min(td, axis=0)
-
-                if D[u, S].size == 0:  # all nodes reached
-                    break
-                minD = np.min(D[u, S])
-                if np.isinf(minD):  # some nodes cannot be reached
-                    break
-                V, = np.where(D[u, :] == minD)
-
-        np.fill_diagonal(D, 1)
-        D = 1 / D
-        np.fill_diagonal(D, 0)
-        return D
-
-    n = len(Gw)
-    Gl = invert(Gw, copy=True)  # connection length matrix
-    A = np.array((Gw != 0), dtype=int)
-    #local efficiency algorithm described by Rubinov and Sporns 2010, not recommended
-    if local == 'original':
-        E = np.zeros((n,))
-        for u in range(n):
-            # V,=np.where(Gw[u,:])		#neighbors
-            # k=len(V)					#degree
-            # if k>=2:					#degree must be at least 2
-            #	e=(distance_inv_wei(Gl[V].T[V])*np.outer(Gw[V,u],Gw[u,V]))**1/3
-            #	E[u]=np.sum(e)/(k*k-k)
-
-            # find pairs of neighbors
-            V, = np.where(np.logical_or(Gw[u, :], Gw[:, u].T))
-            # symmetrized vector of weights
-            sw = cuberoot(Gw[u, V]) + cuberoot(Gw[V, u].T)
-            # inverse distance matrix
-            e = distance_inv_wei(Gl[np.ix_(V, V)])
-            # symmetrized inverse distance matrix
-            se = cuberoot(e) + cuberoot(e.T)
-
-            numer = np.sum(np.outer(sw.T, sw) * se) / 2
-            if numer != 0:
-                # symmetrized adjacency vector
-                sa = A[u, V] + A[V, u].T
-                denom = np.sum(sa)**2 - np.sum(sa * sa)
-                # print numer,denom
-                E[u] = numer / denom  # local efficiency
-
-    #local efficiency algorithm described by Wang et al 2016, recommended
-    elif local in (True, 'local'):
-        E = np.zeros((n,))
-        for u in range(n):
-            V, = np.where(np.logical_or(Gw[u, :], Gw[:, u].T))
-            sw = cuberoot(Gw[u, V]) + cuberoot(Gw[V, u].T)
-            e = distance_inv_wei(cuberoot(Gl)[np.ix_(V, V)])
-            se = e+e.T
-         
-            numer = np.sum(np.outer(sw.T, sw) * se) / 2
-            if numer != 0:
-                # symmetrized adjacency vector
-                sa = A[u, V] + A[V, u].T
-                denom = np.sum(sa)**2 - np.sum(sa * sa)
-                # print numer,denom
-                E[u] = numer / denom  # local efficiency
-
-    elif local in (False, 'global'):
-        e = distance_inv_wei(Gl)
-        E = np.sum(e) / (n * n - n)
-    return E
-
 
 def findpaths(CIJ, qmax, sources, savepths=False):
     '''
@@ -1091,31 +880,3 @@ def mean_first_passage_time(adjacency):
 
     return mfpt
 
-def diffusion_efficiency(adj):
-    '''
-    The diffusion efficiency between nodes i and j is the inverse of the
-    mean first passage time from i to j, that is the expected number of
-    steps it takes a random walker starting at node i to arrive for the
-    first time at node j. Note that the mean first passage time is not a
-    symmetric measure -- mfpt(i,j) may be different from mfpt(j,i) -- and
-    the pair-wise diffusion efficiency matrix is hence also not symmetric.
-
-    Parameters
-    ----------
-    adj : np.ndarray
-        weighted/unweighted, directed/undirected adjacency matrix
-
-    Returns
-    -------
-    gediff : float
-        mean global diffusion efficiency
-    ediff : np.ndarray
-        pairwise NxN diffusion efficiency matrix
-    '''
-    n = len(adj)
-    adj = adj.copy()
-    mfpt = mean_first_passage_time(adj)
-    ediff = 1 / mfpt
-    np.fill_diagonal(ediff, 0)
-    gediff = np.sum(ediff) / (n ** 2 - n)
-    return gediff, ediff
